+++ src/main/java/com/kalsym/locationservice/controller/FeaturedController.java
@@ -102,6 +102,7 @@ public class FeaturedController {
         @RequestParam(required = false) String latitude,
         @RequestParam(required = false) String longitude,
         @RequestParam(required = false) Boolean isMainLevel,
+        @RequestParam(required = false) String storeName,
         @RequestParam(defaultValue = "0") int page,
         @RequestParam(defaultValue = "10") int pageSize,
         @RequestParam(required = false,defaultValue = "sequence") String sortByCol,
@@ -112,7 +113,7 @@ public class FeaturedController {
 
         try{
                     // Page<StoreConfig> body = storeConfigService.getQueryStoreConfig(regionCountryId,cityId,cityName,parentCategoryId,page,pageSize,sortByCol,sortingOrder);
-        Page<StoreConfig> body = storeConfigService.getRawQueryStoreConfig(regionCountryId,cityId,cityName,parentCategoryId,latitude,longitude,searchRadius,isMainLevel,page,pageSize,sortByCol,sortingOrder);
+        Page<StoreConfig> body = storeConfigService.getRawQueryStoreConfig(regionCountryId,cityId,cityName,parentCategoryId,latitude,longitude,searchRadius,isMainLevel,storeName,page,pageSize,sortByCol,sortingOrder);
         response.setData(body);
         response.setStatus(HttpStatus.OK);