+++ src/main/java/com/kalsym/locationservice/model/Config/ProductFeatureConfig.java
@@ -55,5 +55,8 @@ public class ProductFeatureConfig {
     @Column(name="updated_at")
     private Date updatedAt;
 
+    private Integer mainLevelSequence;
+
+
 }