+++ src/main/java/com/kalsym/locationservice/service/StoreConfigService.java
@@ -100,7 +100,7 @@ public class StoreConfigService {
 
     }
 
-    public Page<StoreConfig> getRawQueryStoreConfig(String regionCountryId, List<String> cityId, String cityName, String parentCategoryId,String latitude,String longitude,double searchRadius,Boolean isMainLevel,int page, int pageSize, String sortByCol, Sort.Direction sortingOrder){
+    public Page<StoreConfig> getRawQueryStoreConfig(String regionCountryId, List<String> cityId, String cityName, String parentCategoryId,String latitude,String longitude,double searchRadius,Boolean isMainLevel,String storeName,int page, int pageSize, String sortByCol, Sort.Direction sortingOrder){
     
         StoreConfig storeConfigMatch = new StoreConfig();
 
@@ -121,7 +121,7 @@ public class StoreConfigService {
         .withStringMatcher(ExampleMatcher.StringMatcher.EXACT);
         Example<StoreConfig> example = Example.of(storeConfigMatch, matcher);
 
-        Specification<StoreConfig> storeConfigSpecs = searchStoreConfigSpecs(regionCountryId,cityId, cityName, parentCategoryId,latitude,longitude,searchRadius,isMainLevel,example);
+        Specification<StoreConfig> storeConfigSpecs = searchStoreConfigSpecs(regionCountryId,cityId, cityName, parentCategoryId,latitude,longitude,searchRadius,isMainLevel,storeName,example);
         Page<StoreConfig> result = storeConfigRepository.findAll(storeConfigSpecs, pageable);  
 
         //find the based on location with pageable
@@ -209,6 +209,7 @@ public class StoreConfigService {
         String longitude,
         double radius,
         Boolean isMainLevel,
+        String storeName,
         Example<StoreConfig> example) {
 
         return (Specification<StoreConfig>) (root, query, builder) -> {
@@ -255,6 +256,10 @@ public class StoreConfigService {
 
             if (isMainLevel != null) {
                 predicates.add(builder.equal(root.get("isMainLevel"), isMainLevel));
+            }
+
+            if (storeName != null && !storeName.isEmpty()) {
+                predicates.add(builder.equal(storeDetails.get("name"), storeName));
             }
 
             //use this if you want to group